Use automatic archiver flags
Defining `"rcs"` in the `archiver`'s `RunInfo` breaks with Conan on
MacOS where a prebuilt package is not available for download but has to
be built. The problem is that we configure `ar rcs` as the archiver and
the Conan package's build system again passes similar flags, e.g. `cs`.

The resulting error is as follows:
```
ar: libz.a: No such file or directory
```
As explained in
https://github.com/madler/zlib/issues/588#issuecomment-992694506 .